Added new methods and changed previous ones
add() and remove() both change the size of the array under certain conditions, and append(), first(), last(), and size() were added.  Append() appends to the last empty index of the list, first() returns the first element, last() returns the last element, and size() returns the number of vacant spaces in the list.